[Chore] #533 - 폴더링 및 컨벤션 적용 - 들여쓰기 4칸으로 수정 - ViewController -> VC - 폴더링 다른 Feature와 통일
